Changed the ghost's behaviour for leaving the house
