+++ src/pacman/model/behaviour/GhostLeaveHouseBehaviour.java
@@ -1,7 +1,6 @@
 package pacman.model.behaviour;
 
-import jade.core.Agent;
-import jade.core.behaviours.CyclicBehaviour;
+import jade.core.behaviours.OneShotBehaviour;
 import jade.lang.acl.ACLMessage;
 import java.util.ArrayList;
 import java.util.List;
@@ -11,165 +10,23 @@ import pacman.model.board.Cell;
 import pacman.model.board.CellType;
 import pacman.model.board.Coord2D;
 import pacman.model.board.Direction;
-import pacman.model.core.Constant;
-import pacman.model.core.GameVocabulary;
 import pacman.model.core.GhostVocabulary;
 
-public class GhostLeaveHouseBehaviour extends CyclicBehaviour
+public abstract class GhostLeaveHouseBehaviour extends OneShotBehaviour
 {
 
-    private final Board board;
-    private final Cell myCell;
+    protected  final Board board;
+    protected  final Cell myCell;
 
-    public GhostLeaveHouseBehaviour(Agent agent, Board board, Cell cell)
+    public GhostLeaveHouseBehaviour(Board board, Cell cell)
     {
-        super(agent);
+        //super(agent);
         this.board = board;
         this.myCell = cell;
     }
-
-    @Override
-    public void action()
-    {
-        ACLMessage message = myAgent.receive();
-        if (null != message && ACLMessage.INFORM == message.getPerformative())
-        {
-            System.out.println(myAgent.getLocalName() + " received " + message.getContent() + " from " + message.getSender().getLocalName());
-            
-            switch (message.getOntology())
-            {
-                case GameVocabulary.ONTOLOGY:
-                    switch (message.getContent())
-                    {
-                        case GameVocabulary.START:
-                            firstGhostToLeave();
-                            break;
-                            
-                        default:
-                            block();
-                    }
-                    
-                    break;
-                    
-                case GhostVocabulary.ONTOLOGY:
-                    switch (message.getContent())
-                    {
-                        case GhostVocabulary.LEAVE_THE_HOUSE:
-                            nextGhostToLeave();
-                            break;
-                            
-                        default:
-                            block();
-                    }
-                    
-                default:
-                    block();
-            }
-        } 
-        else
-        {
-            block();
-        }
-    }
-    
-    private void firstGhostToLeave()
-    {
-        // Delays a time before the first ghost leaves his house
-        try { Thread.sleep(Constant.MOVEMENT_DELAY); } catch (InterruptedException ex) {}
-                        
-        // Tries to leave the house
-        Coord2D doorPosition = board.getGhostDoor().getPosition();
-        Coord2D myNewPosition = null;
-        int x = myCell.getPosition().x,
-            y = myCell.getPosition().y;
-
-        for (Direction direction : Direction.values())
-        {
-            if (doorPosition.equals(myNewPosition = new Coord2D(x + direction.xInc, y + direction.yInc)))
-            {
-                System.out.println(myAgent.getLocalName() + " leaving house by " + direction.toString().toLowerCase());
-                break;
-            }
-
-            myNewPosition = null; // Resets the position to null so the others ghosts don't try to use it
-        }
-
-        if (null != myNewPosition)
-        {
-            // Leaves the house
-            ((GhostAgent) myAgent).setHouseLeft(true);
-            board.moveCell(myCell, myNewPosition);
-            board.print();
-            
-            // Advices the next ghost (if any) to leave his house
-            adviceNextGhost();
-            
-            // Delays a time before the first ghost leaves the door
-            try { Thread.sleep(Constant.MOVEMENT_DELAY); } catch (InterruptedException ex) {}
-        }
-    }
     
     @SuppressWarnings("SleepWhileInLoop")
-    private void nextGhostToLeave()
-    {
-        // Delays a time and starts to move
-        try { Thread.sleep(Constant.GHOST_LEAVE_HOUSE_DELAY); } catch (InterruptedException ex) {}
-                    
-        Coord2D doorPosition = board.getGhostDoor().getPosition();
-        Direction lastDirection = null;
-        
-        do 
-        {
-            for (Direction direction : Direction.values())
-            {
-                Coord2D myPosition = myCell.getPosition();
-                Coord2D myNewPosition = new Coord2D(myPosition.x + direction.xInc, myPosition.y + direction.yInc);
-                        
-                // Found the door
-                if (doorPosition.equals(myNewPosition))
-                {
-                    // Leaves the house
-                    System.out.println(myAgent.getLocalName() + " leaving house by " + direction.toString().toLowerCase());
-                    ((GhostAgent) myAgent).setHouseLeft(true);
-                    board.moveCell(myCell, myNewPosition);
-                    board.print();
-            
-                    // Advices the next ghost (if any) to leave his house
-                    adviceNextGhost();
-                    
-                    // Delays a time before the next ghost leaves the door
-                    try { Thread.sleep(Constant.MOVEMENT_DELAY); } catch (InterruptedException ex) {}
-            
-                    break;
-                }
-                // Else, keeps looking for it
-                else
-                {
-                    Cell nearCell = board.getCell(myNewPosition);
-                    if (CellType.GHOST_HOUSE == nearCell.getType())
-                    {
-                        if (null == lastDirection)
-                        {
-                            lastDirection = direction;
-                        }
-
-                        if (lastDirection == direction)
-                        {
-                            System.out.println(myAgent.getLocalName() + " is moving to " + direction);   
-                            board.moveCell(myCell, myNewPosition);
-                            board.print();    
-
-                            // Delays a time after the ghost makes its movement
-                            try { Thread.sleep(Constant.MOVEMENT_DELAY); } catch (InterruptedException ex) {}
-                        }
-                    }
-                }
-            }
-        } while (!((GhostAgent) myAgent).isHouseLeft());
-    }
-    
-    @SuppressWarnings("SleepWhileInLoop")
-    private void adviceNextGhost()
+    protected void adviceNextGhost()
     {
         // Selects the next ghost (if any) to leave the house and sends it a message
         List<GhostAgent> nextGhosts = new ArrayList<>(); // Just for a hack
